removed outcomefilter and filefilter
