--- frontend/src/@types/FileFilter.d.ts
@@ -1,7 +1,0 @@
-import { TransactionLogFilter } from "./TransactionLogFilter";
-
-export interface FileFilters {
-    id: string,
-    filterName: string,
-    checkboxList: Array<TransactionLogFilter>
-};
