--- frontend/src/@types/TransactionLogFilter.d.ts
@@ -1,8 +1,0 @@
-export interface TransactionLogFilter {
-    id: string,
-    name?: string,
-    format?: string,
-    isChecked: boolean,
-    value?: string,
-    filter?: string  
-};
