+++ frontend/src/types/TransactionLogFilter.d.ts
@@ -1,0 +1,16 @@
+export interface TransactionLogFilter {
+    id: string,
+    name?: string,
+    format?: string,
+    isChecked: boolean,
+    value?: string,
+    filter?: string,
+    type?: string,
+    titleColor?: string
+};
+
+export interface TransactionLogFilters {
+    id: string,
+    filterName: string,
+    checkboxList: Array<TransactionLogFilter>
+};
